+++ ebzl/modules/create.py
@@ -13,7 +13,7 @@ import boto
 def get_argument_parser():
     parser = argparse.ArgumentParser("ebzl create")
 
-    parameters.add_profile(parser)
+    parameters.add_profile(parser, required=False)
     parameters.add_app_name(parser)
     parameters.add_version_label(parser, required=False)
     parameters.add_region(parser, required=False)