+++ ebzl/modules/deploy.py
@@ -10,7 +10,7 @@ import argparse
 def get_argument_parser():
     parser = argparse.ArgumentParser("ebzl deploy")
 
-    parameters.add_profile(parser)
+    parameters.add_profile(parser, required=False)
     parameters.add_version_label(parser, required=False)
     parameters.add_region(parser, required=False)
     parameters.add_env_name(parser, required=True)